% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildFooter.R
\name{buildFooter}
\alias{buildFooter}
\title{show df's missing ratio, number of variables and column types..}
\usage{
buildFooter(title = "Data Visualization App", version = 1, subversion = 1, URL)
}
\arguments{
\item{URL}{various ~~~~.}

\item{x}{version}

\item{y}{subversion}
}
\description{
show df's missing ratio, number of variables and column types..
}
\examples{
describe_df(billboard)
}
\author{
Byungju Kim (bjk@inu.ac.kr)
}
