% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildPage.R
\name{buildPage}
\alias{buildPage}
\title{Page builder for Shiny App with Appsilon's Shiny fluent.}
\usage{
buildPage(title, subtitle, contents)
}
\arguments{
\item{title}{title of the page}

\item{subtitle}{subtitle of the page}

\item{contents}{various html codes required to build page.}
}
\description{
shamlessly stolen from Appsilon's tutorial.
}
\examples{
buildPage()
}
\author{
Byungju Kim (bjk@inu.ac.kr)
}
