% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callNickname.R
\name{callNickname}
\alias{callNickname}
\title{Redirect version nickname based on 10-chenkan, 12-zodiac system.}
\usage{
callNickname(x = 0, y = 0)
}
\arguments{
\item{x}{version}

\item{y}{subversion}
}
\description{
Redirect version nickname based on 10-chenkan, 12-zodiac system.
}
\examples{
BKversion(1,1)
}
\author{
Byungju Kim (bjk@inu.ac.kr)
}
