% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daviz.R
\name{daviz}
\alias{daviz}
\title{show df's missing ratio, number of variables and column types..}
\usage{
daviz(meta, data, ...)
}
\arguments{
\item{meta}{a data.frame of tibble object for meta data.}

\item{data}{a data.frame or tibble object.}

\item{...}{various ~~~~.}
}
\description{
show df's missing ratio, number of variables and column types..
}
\examples{
describe_df(billboard)
}
\author{
Byungju Kim (bjk@inu.ac.kr)
}
