% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daviz_version.R
\name{daviz_version}
\alias{daviz_version}
\title{show df's missing ratio, number of variables and column types..}
\usage{
daviz_version(
  title = "DaViz",
  ver = version,
  subv = subversion,
  AUTHOR = "KIM, BYUNGJU PhD.",
  url = "https://byungju-kim.github.io/DaViz/"
)
}
\arguments{
\item{title}{software title, "DaViz".}

\item{AUTHOR}{author of the program.}

\item{url}{url for the software}

\item{version}{official version.}

\item{subversion}{official subversion.}
}
\description{
show df's missing ratio, number of variables and column types..
}
\examples{
daviz_version()
}
\author{
Byungju Kim (bjk@inu.ac.kr)
}
