% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dec2base.R
\name{dec2base}
\alias{dec2base}
\title{convert Decimal numners to each of base 'n' system.}
\usage{
dec2base(x, base, format = T)
}
\arguments{
\item{x}{a vector of decimal numeric values}

\item{base}{a base number which you want to convert to}

\item{format}{a logical value (T/F) which formating leading zero based on most long number.}
}
\description{
convert Decimal numners to each of base 'n' system.
}
\examples{
dec2base(1:10,3,T)
}
\author{
Byungju Kim (bjk@inu.ac.kr)
}
