% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe_df.R
\name{describe_df}
\alias{describe_df}
\title{show df's missing ratio, number of variables and column types..}
\usage{
describe_df(x, desc, n_var = 7, show_freq = F)
}
\arguments{
\item{x}{a data.frame or tibble object.}

\item{desc}{column description file (under-development)}

\item{n_var}{choice of one-hot encoding or additive model.}
}
\description{
show df's missing ratio, number of variables and column types..
}
\examples{
describe_df(billboard)
}
\author{
Byungju Kim (bjk@inu.ac.kr)
}
