% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gauge_df.R
\name{gauge}
\alias{gauge}
\title{show df's mossing ratio, number of variables and column types..}
\usage{
gauge(x, desc, n_var = 7, show_freq = F)
}
\arguments{
\item{x}{a data.frame or tibble object.}

\item{desc}{column description file (under-development)}

\item{n_var}{choice of one-hot encoding or additive model.}
}
\description{
show df's mossing ratio, number of variables and column types..
}
\examples{
gauge(billboard)
}
\author{
Byungju Kim (bjk@inu.ac.kr)
}
