% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/request.R
\name{is.installed}
\alias{is.installed}
\title{check the package installation}
\usage{
is.installed(pkg)
}
\arguments{
\item{mypkg}{character vector of package names interested.}
}
\description{
check if the package is installed
}
\examples{
is.installed("tidyverse")
}
