% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ped2Add.R
\name{ped2Add}
\alias{ped2Add}
\title{Read plink 12 coding files}
\usage{
ped2Add(ped, map, onehot = F)
}
\arguments{
\item{ped}{a 1/2 coded plnk pedigree file (+7 metadata column)}

\item{map}{plink genomic map file which describe genomic coordinate and unique SNP IDs.}
}
\description{
Read plink 12 coding files
}
\examples{
ped2Add(ped,map)

}
\author{
Byungju Kim (bjk@inu.ac.kr)
}
