% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eigenplots.R
\name{plot_eval}
\alias{plot_eval}
\title{eigenvalue plot including % cumulative eigenvalue}
\usage{
plot_eval(eval, PCs)
}
\description{
eigenvalue plot including % cumulative eigenvalue
}
\examples{
PCA<-prcomp(iris[,1:4])
plot_eval(PCA$sdev^2,3)
}
