% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eigenplots.R
\name{plot_evec}
\alias{plot_evec}
\title{create various PCA graphes}
\usage{
plot_evec(evec, axis, color, shape, colors, shapes)
}
\arguments{
\item{x}{vector, and list objects.}
}
\description{
create various PCA graphes
}
\examples{
BKtbl(billboard)
}
\author{
Byungju Kim (bjk@inu.ac.kr)
}
