% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_eval.R
\name{plote_eval}
\alias{plote_eval}
\alias{plot_eval}
\title{eigenvalue plot including % cumulative eigenvalue}
\usage{
plot_eval(eval, PCs)
}
\arguments{
\item{eval}{vector of eigenvalues}

\item{PCs}{number of Principle components to draw chart}
}
\description{
eigenvalue plot including % cumulative eigenvalue
}
\examples{
PCA<-prcomp(iris[,1:4])
plot_eval(PCA$sdev^2,3)
}
\author{
Byungju Kim (bjk@inu.ac.kr)
}
