% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/request.R
\name{Q.library}
\alias{Q.library}
\title{load package quietly}
\usage{
Q.library(pkg)
}
\arguments{
\item{pkg}{packages of interests.}
}
\description{
wrapper function for library
}
\examples{
Q.library("tidyverse")
}
