% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/request.R
\name{Q.Require}
\alias{Q.Require}
\title{require package quietly}
\usage{
Q.Require(pkg)
}
\arguments{
\item{pkg}{packages of interests.}
}
\description{
wrapper function for require
}
\examples{
Q.Require("tidyverse")
}
