% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.dist.R
\name{read.dist}
\alias{read.dist}
\alias{read.phylip_dist}
\title{read.dist: read Phylip style distance matrix as distance matrix in R}
\usage{
read.phylip_dist(file, rm.list = c())
}
\arguments{
\item{file}{a file name,}

\item{rm.list}{list for sample exclusion.}

\item{skip}{a numeric valueof line skipping.}
}
\description{
read.dist: read Phylip style distance matrix as distance matrix in R
}
\examples{
 read.dist(matrix.tsv)
}
\author{
Byungju Kim (bjk@inu.ac.kr)
}
