% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/request.R
\name{request}
\alias{request}
\title{install and load multiple packages}
\usage{
request(pkgs)
}
\arguments{
\item{pkgs}{vector of packages interested. if installation is necessary, do installation first from cran.}
}
\description{
if installation is necessary, do installation first from cran and then load all the packages queitly.
}
