% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shuf.dist.R
\name{shuf.dist}
\alias{shuf.dist}
\title{Random shuffling of distance matrix.}
\usage{
shuf.dist(x, seed = 716018)
}
\arguments{
\item{x}{a dist object.}

\item{seed}{seed used to generate random sampling order.}
}
\description{
shuffle distance matrix randomly using a given seed
}
\examples{
shuf.dist(x,seed=1234)
}
\author{
Byungju Kim (bjk@inu.ac.kr)
}
