% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildFooter.R
\name{buildFooter}
\alias{buildFooter}
\title{footer builder for DaViz App.}
\usage{
buildFooter(title = "Data Visualization App", version = 1, subversion = 1, URL)
}
\arguments{
\item{URL}{URL to add.}

\item{x}{version}

\item{y}{subversion}
}
\description{
show df's missing ratio, number of variables and column types..
}
\examples{
buildFooter(billboard)
}
\author{
Byungju Kim (bjk@inu.ac.kr)
}
