% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildPage.R
\name{buildPage}
\alias{buildPage}
\title{Page builder for Shiny App with Appsilon's Shiny fluent.}
\usage{
buildPage(title, subtitle, contents)
}
\arguments{
\item{x}{version}

\item{y}{subversion}

\item{URL}{various ~~~~.}
}
\description{
Page builder for Shiny App with Appsilon's Shiny fluent.
}
\examples{
describe_df(billboard)
}
\author{
Byungju Kim (bjk@inu.ac.kr)
}
