% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daviz.R
\name{daviz}
\alias{daviz}
\title{Interactive Exploratory data analysis on data.}
\usage{
daviz(meta, data, ...)
}
\arguments{
\item{meta}{a data.frame of tibble object for meta data.}

\item{data}{a data.frame or tibble object.}

\item{...}{various ~~~~.}
}
\description{
Interactive Exploratory data analysis on data using daviz and data handler.
}
\examples{
daviz(mtcars,mtcars)
}
\author{
Byungju Kim (bjk@inu.ac.kr)
}
