% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_eval.R
\name{plot_eval}
\alias{plot_eval}
\title{functions for cumulative eigenvalue and eigenvalue plots}
\usage{
plot_eval(eval, PCs)
}
\arguments{
\item{eval}{vector of eigenvalues}

\item{PCs}{number of Principle components to draw chart}
}
\description{
functions for cumulative eigenvalue and eigenvalue plots
}
\examples{
PCA<-prcomp(iris[,1:4])
plot_eval(PCA$sdev^2,3)
}
\author{
Byungju Kim (bjk@inu.ac.kr)
}
